% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixData.R
\docType{methods}
\name{[,SaemixData-method}
\alias{[,SaemixData-method}
\title{Get/set methods for SaemixData object}
\usage{
\S4method{[}{SaemixData}(x, i, j, drop)
}
\arguments{
\item{x}{object}

\item{i}{element to be replaced}

\item{j}{element to replace with}

\item{drop}{whether to drop unused dimensions}
}
\description{
Access slots of a SaemixData object using the object["slot"] format
}
\keyword{methods}
