% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixObject.R
\docType{methods}
\name{[,SaemixObject-method}
\alias{[,SaemixObject-method}
\title{Get/set methods for SaemixObject object}
\usage{
\S4method{[}{SaemixObject}(x, i, j, drop)
}
\arguments{
\item{x}{object}

\item{i}{element to be replaced}

\item{j}{element to replace with}

\item{drop}{whether to drop unused dimensions}
}
\description{
Access slots of a SaemixObject object using the object["slot"] format
}
\keyword{methods}
